# other/api/swagger.yaml
swagger: "2.0"
info:
  description: "This is a simple bookstore server with a book inventory, users and a shopping cart."
  version: "1.0.0"
  title: "Simple Bookstore"
  contact:
    email: "10451768@polimi.it"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /user/login:
    post:
      tags:
      - "user"
      summary: "Login"
      description: "Login with a form"
      operationId: "userLoginPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "email"
        in: "formData"
        required: true
        type: "string"
      - name: "password"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "succesfull login"
        400:
          description: "bad request"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /user/logout:
    delete:
      tags:
      - "user"
      summary: "Logout"
      description: "Logout, delete session"
      operationId: "userLogoutDELETE"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      responses:
        200:
          description: "succesfull logout"
        400:
          description: "bad request"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /user/register:
    post:
      tags:
      - "user"
      summary: "Register"
      description: "Register into the store"
      operationId: "userRegisterPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "succesfull registration"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /cart:
    get:
      tags:
      - "cart"
      summary: "Get the content of the cart"
      operationId: "getCartById"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 4 and cannot exceed 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "succesful operartion"
          schema:
            $ref: "#/definitions/Cart"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "Cart"
    post:
      tags:
      - "cart"
      summary: "Add a new book to the cart"
      operationId: "addBookById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "book"
        description: "Book object that needs to be added to the cart"
        required: true
        schema:
          $ref: "#/definitions/Cart"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Cart"
  /cart/count:
    get:
      tags:
      - "cart"
      summary: "Get the number of books in cart"
      operationId: "getCartCountById"
      produces:
      - "application/json"
      responses:
        200:
          description: "succesful operartion"
          schema:
            $ref: "#/definitions/Cart"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "Cart"
  /cart/{Id_book}:
    put:
      tags:
      - "cart"
      summary: "Edit the quantity of a book in the cart"
      operationId: "updateBookById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id_book"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Book object that needs to be modified to the cart"
        required: true
        schema:
          $ref: "#/definitions/Cart"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Cart"
    delete:
      tags:
      - "cart"
      summary: "Delete a book in the cart"
      operationId: "deleteBookById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id_book"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Cart"
...
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
    title: "User"
    description: "A user"
    example:
      id: 1
      name: "Giacomo"
      address: "DEIB"
      creditcard: "xyzabc"
...
  Cart:
    type: "object"
    properties:
      total:
        $ref: "#/definitions/Amount"
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"
    title: "Cart"
    description: "Order for books"
    example:
      total:
        currency: "eur"
        value: 80082819046101.171875
      books:
      - id: 0
        title: "Il deserto dei tartari"
        author: "Dino Buzzati"
        price:
          value: 10
          currency: "eur"
      - id: 0
        title: "Il deserto dei tartari"
        author: "Dino Buzzati"
        price:
          value: 10
          currency: "eur"
  Book:
    type: "object"
    required:
    - "price"
    - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
      id_genre:
        type: "integer"
        format: "int64"
      id_themes:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      description:
        type: "string"
      cover_img:
        type: "string"
      pages:
        type: "string"
      rating:
        type: "integer"
      price:
        $ref: "#/definitions/Amount"
      status:
        type: "string"
        description: "book availability in the inventory"
        enum:
        - "available"
        - "out of stock"
      support:
        type: "string"
        description: "book support in the inventory"
        enum:
        - "paper_based"
        - "ebook"
        - "both"
    title: "Book"
    description: "A book for sale in the store"
    example:
      id: 0
      title: "Il deserto dei tartari"
      author: "Dino Buzzati"
      price:
        value: 10
        currency: "eur"
      status: "available"
...
  Amount:
    type: "object"
    required:
    - "currency"
    - "value"
    properties:
      value:
        type: "number"
        format: "double"
        minimum: 0.01
        maximum: 1000000000000000
      currency:
        $ref: "#/definitions/Currency"
    description: "Price\n"
    example:
      currency: "eur"
      value: 11
  Currency:
    type: "string"
    pattern: "^[A-Z]{3,3}$"
    description: "some description\n"
    example: "eur"
...